refactor: Operations, Details
